requirements.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.9.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
